# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مقرأة سَنا الآي Quran Center Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shortcomings:
    get:
      operationId: listShortcomings
      tags: [shortcomings]
      summary: List sessions with shortcomings (تقصير)
      parameters:
        - name: circleId
          in: query
          schema:
            type: integer
        - name: trackName
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of shortcoming records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShortcomingItem"
  /records/{id}/shortcoming:
    patch:
      operationId: updateShortcomingOverride
      tags: [shortcomings]
      summary: Override shortcoming flag for a record (track_supervisor within 48h)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shortcomingOverride:
                  type: ["boolean", "null"]
              required: [shortcomingOverride]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortcomingItem"
        "403":
          description: Not allowed
        "404":
          description: Record not found
...
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a student, circle, or track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    get:
      operationId: listMessages
      tags: [messages]
      summary: List all sent messages (leader only)
      responses:
        "200":
          description: All messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/batch:
    post:
      operationId: batchSendMessages
      tags: [messages]
      summary: Send the same message to multiple students at once (leader only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchSendMessageBody"
      responses:
        "201":
          description: Messages sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                required: [count]
...
  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message (leader only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Message deleted
components:
  schemas:
...
    ShortcomingItem:
      type: object
      properties:
        recordId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        circleId:
          type: integer
        circleName:
          type: string
        trackName:
          type: string
        date:
          type: string
        reasons:
          type: array
          items:
            type: string
        shortcomingOverride:
          type: ["boolean", "null"]
        isShortcoming:
          type: boolean
        canEdit:
          type: boolean
        createdAt:
          type: string
      required:
        - recordId
        - studentId
        - studentName
        - circleId
        - circleName
        - trackName
        - date
        - reasons
        - isShortcoming
        - canEdit
        - createdAt
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        targetType:
          type: string
        targetId:
          type: string
        targetLabel:
          type: string
        content:
          type: string
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, senderId, senderName, targetType, targetId, targetLabel, content, createdAt]
    SendMessageBody:
      type: object
      properties:
        targetType:
          type: string
        targetId:
          type: string
        content:
          type: string
        expiresAt:
          type: ["string", "null"]
      required: [targetType, targetId, content]
...
    BatchSendMessageBody:
      type: object
      properties:
        studentIds:
          type: array
          items:
            type: integer
        content:
          type: string
        expiresAt:
          type: ["string", "null"]
      required: [studentIds, content]
...